move audit fix tests into their respective test files
