+++ tests/test_body_digest.py
@@ -1,5 +1,7 @@
 """Tests for body digest computation and verification."""
 
+import json
+
 from mpp._body_digest import compute, verify
 
 
@@ -64,6 +66,24 @@ class TestCompute:
         payload = digest.removeprefix("sha-256=")
         raw = base64.b64decode(payload)
         assert len(raw) == 32
+
+    def test_key_order_independent(self) -> None:
+        """Dict digests should be identical regardless of key insertion order."""
+        d1 = {"z": "1", "a": "2", "m": "3"}
+        d2 = {"a": "2", "m": "3", "z": "1"}
+        assert compute(d1) == compute(d2)
+
+    def test_matches_canonical_json(self) -> None:
+        """Dict digest should match digest of canonical (sorted) JSON string."""
+        d = {"b": "2", "a": "1"}
+        canonical = json.dumps(d, separators=(",", ":"), sort_keys=True, ensure_ascii=False)
+        assert compute(d) == compute(canonical)
+
+    def test_nested_dict_key_order(self) -> None:
+        """Nested dict should also be sorted by keys."""
+        d1 = {"outer_z": {"inner_b": 1, "inner_a": 2}, "outer_a": 3}
+        d2 = {"outer_a": 3, "outer_z": {"inner_a": 2, "inner_b": 1}}
+        assert compute(d1) == compute(d2)
 
 
 class TestVerify:
