+++ tests/test_mcp.py
@@ -1,5 +1,7 @@
 """Tests for MCP transport support."""
 
+import base64
+import json
 from datetime import UTC, datetime
 
 import pytest
@@ -36,6 +38,8 @@ def _make_bound_mcp_challenge(
     secret_key: str = MCP_TEST_SECRET,
     expires: str | None = None,
     description: str | None = None,
+    digest: str | None = None,
+    opaque: dict[str, str] | None = None,
 ) -> MCPChallenge:
     """Create an MCPChallenge with an HMAC-bound ID for testing."""
     if request is None:
@@ -47,6 +51,8 @@ def _make_bound_mcp_challenge(
         intent=intent,
         request=request,
         expires=expires,
+        digest=digest,
+        opaque=opaque,
     )
     return MCPChallenge(
         id=challenge_id,
@@ -56,6 +62,8 @@ def _make_bound_mcp_challenge(
         request=request,
         expires=expires,
         description=description,
+        digest=digest,
+        opaque=opaque,
     )
 
 
@@ -754,3 +762,253 @@ class TestCreateChallenge:
         # But same secret+params should produce consistent HMACs
         assert len(challenge1.id) > 0
         assert len(challenge2.id) > 0
+
+
+class TestMCPDigestOpaque:
+    """PY-01: MCPChallenge digest and opaque field handling."""
+
+    def test_mcp_challenge_has_digest_and_opaque_fields(self) -> None:
+        ch = MCPChallenge(
+            id="test",
+            realm="r",
+            method="tempo",
+            intent="charge",
+            request={"amount": "1"},
+            digest="sha-256=abc",
+            opaque={"pi": "pi_123"},
+        )
+        assert ch.digest == "sha-256=abc"
+        assert ch.opaque == {"pi": "pi_123"}
+
+    def test_mcp_challenge_to_dict_includes_digest_opaque(self) -> None:
+        ch = MCPChallenge(
+            id="test",
+            realm="r",
+            method="tempo",
+            intent="charge",
+            request={},
+            digest="sha-256=xyz",
+            opaque={"k": "v"},
+        )
+        d = ch.to_dict()
+        assert d["digest"] == "sha-256=xyz"
+        assert d["opaque"] == {"k": "v"}
+
+    def test_mcp_challenge_to_dict_omits_none_digest_opaque(self) -> None:
+        ch = MCPChallenge(id="test", realm="r", method="tempo", intent="charge", request={})
+        d = ch.to_dict()
+        assert "digest" not in d
+        assert "opaque" not in d
+
+    def test_mcp_challenge_from_dict_parses_digest_opaque(self) -> None:
+        data = {
+            "id": "test",
+            "realm": "r",
+            "method": "tempo",
+            "intent": "charge",
+            "request": {},
+            "digest": "sha-256=abc",
+            "opaque": {"k": "v"},
+        }
+        ch = MCPChallenge.from_dict(data)
+        assert ch.digest == "sha-256=abc"
+        assert ch.opaque == {"k": "v"}
+
+    @pytest.mark.asyncio
+    async def test_mcp_verify_succeeds_with_digest(self) -> None:
+        class MockIntent:
+            name = "charge"
+
+            async def verify(self, credential: object, request: dict) -> Receipt:
+                return Receipt.success(reference="0x123")
+
+        request = {"amount": "1000"}
+        digest_val = "sha-256=abc123"
+        challenge = _make_bound_mcp_challenge(
+            request=request,
+            digest=digest_val,
+        )
+        cred = MCPCredential(challenge=challenge, payload={"sig": "0x"})
+
+        result = await verify_or_challenge(
+            meta=cred.to_meta(),
+            intent=MockIntent(),
+            request=request,
+            realm="api.example.com",
+            secret_key=MCP_TEST_SECRET,
+        )
+        assert isinstance(result, tuple), "Should verify successfully with digest"
+
+    @pytest.mark.asyncio
+    async def test_mcp_verify_succeeds_with_opaque(self) -> None:
+        class MockIntent:
+            name = "charge"
+
+            async def verify(self, credential: object, request: dict) -> Receipt:
+                return Receipt.success(reference="0x123")
+
+        request = {"amount": "1000"}
+        opaque = {"pi": "pi_abc"}
+        challenge = _make_bound_mcp_challenge(
+            request=request,
+            opaque=opaque,
+        )
+        cred = MCPCredential(challenge=challenge, payload={"sig": "0x"})
+
+        result = await verify_or_challenge(
+            meta=cred.to_meta(),
+            intent=MockIntent(),
+            request=request,
+            realm="api.example.com",
+            secret_key=MCP_TEST_SECRET,
+        )
+        assert isinstance(result, tuple), "Should verify successfully with opaque"
+
+    @pytest.mark.asyncio
+    async def test_mcp_verify_fails_without_matching_digest(self) -> None:
+        class MockIntent:
+            name = "charge"
+
+            async def verify(self, credential: object, request: dict) -> Receipt:
+                return Receipt.success(reference="0x123")
+
+        request = {"amount": "1000"}
+        challenge_id = generate_challenge_id(
+            secret_key=MCP_TEST_SECRET,
+            realm="api.example.com",
+            method="tempo",
+            intent="charge",
+            request=request,
+            digest="sha-256=original",
+        )
+        ch = MCPChallenge(
+            id=challenge_id,
+            realm="api.example.com",
+            method="tempo",
+            intent="charge",
+            request=request,
+            digest="sha-256=tampered",
+        )
+        cred = MCPCredential(challenge=ch, payload={"sig": "0x"})
+
+        result = await verify_or_challenge(
+            meta=cred.to_meta(),
+            intent=MockIntent(),
+            request=request,
+            realm="api.example.com",
+            secret_key=MCP_TEST_SECRET,
+        )
+        assert isinstance(result, MCPChallenge)
+
+
+class TestMCPCredentialToCoreDigestOpaque:
+    """PY-06: MCPCredential.to_core() must pass digest/opaque to ChallengeEcho."""
+
+    def test_to_core_includes_digest(self) -> None:
+        ch = MCPChallenge(
+            id="test",
+            realm="r",
+            method="tempo",
+            intent="charge",
+            request={"amount": "1"},
+            digest="sha-256=abc123",
+        )
+        cred = MCPCredential(challenge=ch, payload={"sig": "0x"})
+        core = cred.to_core()
+        assert core.challenge.digest == "sha-256=abc123"
+
+    def test_to_core_includes_opaque(self) -> None:
+        ch = MCPChallenge(
+            id="test",
+            realm="r",
+            method="tempo",
+            intent="charge",
+            request={"amount": "1"},
+            opaque={"pi": "pi_123"},
+        )
+        cred = MCPCredential(challenge=ch, payload={"sig": "0x"})
+        core = cred.to_core()
+        assert core.challenge.opaque is not None
+        decoded = json.loads(base64.urlsafe_b64decode(core.challenge.opaque + "==").decode())
+        assert decoded == {"pi": "pi_123"}
+
+    def test_to_core_opaque_is_sorted(self) -> None:
+        ch = MCPChallenge(
+            id="test",
+            realm="r",
+            method="tempo",
+            intent="charge",
+            request={},
+            opaque={"z": "last", "a": "first"},
+        )
+        cred = MCPCredential(challenge=ch, payload={})
+        core = cred.to_core()
+        assert core.challenge.opaque is not None
+        decoded_json = base64.urlsafe_b64decode(core.challenge.opaque + "==").decode()
+        assert decoded_json.index('"a"') < decoded_json.index('"z"')
+
+    def test_to_core_none_digest_opaque(self) -> None:
+        ch = MCPChallenge(
+            id="test",
+            realm="r",
+            method="tempo",
+            intent="charge",
+            request={"amount": "1"},
+        )
+        cred = MCPCredential(challenge=ch, payload={"sig": "0x"})
+        core = cred.to_core()
+        assert core.challenge.digest is None
+        assert core.challenge.opaque is None
+
+    def test_to_core_roundtrip_hmac_with_opaque(self) -> None:
+        opaque = {"pi": "pi_abc"}
+        request = {"amount": "1000"}
+        challenge = _make_bound_mcp_challenge(
+            request=request,
+            opaque=opaque,
+        )
+        cred = MCPCredential(challenge=challenge, payload={"sig": "0x"})
+        core = cred.to_core()
+
+        from mpp._parsing import _b64_decode
+
+        echo = core.challenge
+        echo_request = _b64_decode(echo.request) if echo.request else {}
+        echo_opaque = _b64_decode(echo.opaque) if echo.opaque else None
+        recomputed_id = generate_challenge_id(
+            secret_key=MCP_TEST_SECRET,
+            realm=echo.realm,
+            method=echo.method,
+            intent=echo.intent,
+            request=echo_request,
+            expires=echo.expires,
+            digest=echo.digest,
+            opaque=echo_opaque,
+        )
+        assert recomputed_id == challenge.id
+
+    def test_to_core_roundtrip_hmac_with_digest(self) -> None:
+        request = {"amount": "1000"}
+        digest_val = "sha-256=test123"
+        challenge = _make_bound_mcp_challenge(
+            request=request,
+            digest=digest_val,
+        )
+        cred = MCPCredential(challenge=challenge, payload={"sig": "0x"})
+        core = cred.to_core()
+
+        from mpp._parsing import _b64_decode
+
+        echo = core.challenge
+        echo_request = _b64_decode(echo.request) if echo.request else {}
+        recomputed_id = generate_challenge_id(
+            secret_key=MCP_TEST_SECRET,
+            realm=echo.realm,
+            method=echo.method,
+            intent=echo.intent,
+            request=echo_request,
+            expires=echo.expires,
+            digest=echo.digest,
+            opaque=None,
+        )
+        assert recomputed_id == challenge.id
